chore: rename lib/platform/dfwhook.h => lib/platform/Win32Hook.h
