+++ src/lib/platform/MSWindowsDesks.h
@@ -13,7 +13,7 @@
 #include "deskflow/OptionTypes.h"
 #include "mt/CondVar.h"
 #include "mt/Mutex.h"
-#include "platform/dfwhook.h"
+#include "platform/Win32Hook.h"
 
 #include <string>
