+++ src/lib/platform/Win32Hook.h
